refactor: retrieve routes by function rather than constant
