+++ utils/routes.ts
@@ -1,6 +1,15 @@
-export const ROUTES = [
+export type Route = {
+  text: string;
+  href: string;
+};
+
+const ROUTES: Route[] = [
   { text: 'The Oca Lab', href: '/' },
   { text: 'Teaching', href: '/courses' },
   { text: 'News', href: '/' },
   { text: 'Team', href: '/' },
-] as const;
+];
+
+export async function retrieveRoutes() {
+  return ROUTES;
+}
